requirements.txt
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
html2text==2024.2.26
langchain==0.3.14
langchain-aws==0.2.11
langchain-core==0.3.29
langchain-ollama==0.2.2
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
flask==3.0.2
tabulate==0.9.0